rpm: dnf-phytec-config: Add initial config
Add an initial configuration for dnf and a systemd service for creating
the cache. The service is only started once for the first boot, if the
example repo file still exists. We need to copy the repo file to its
final place on the running system. If we install it during build-time,
Bitbake scans it for installing packages needed for the build, but this
is prone to fail and unwanted, as the package repository may not be
available then.

Signed-off-by: Martin Schwan <m.schwan@phytec.de>